requirements.txt
easyocr==1.5.0
matplotlib==3.2.1
numpy==1.19.4
opencv_contrib_python==4.5.1.48
Pillow==9.1.1
streamlit==1.5.0
